--- a/core/config/project_settings.cpp
+++ b/core/config/project_settings.cpp
@@ -39,7 +39,6 @@
 #include "core/io/file_access_pack.h"
 #include "core/io/marshalls.h"
 #include "core/os/keyboard.h"
-#include "core/os/os.h"
 #include "core/variant/variant_parser.h"
 #include "core/version.h"