--- a/core/core_bind.cpp
+++ b/core/core_bind.cpp
@@ -39,7 +39,6 @@
 #include "core/math/geometry_2d.h"
 #include "core/math/geometry_3d.h"
 #include "core/os/keyboard.h"
-#include "core/os/os.h"
 
 namespace core_bind {