--- a/core/os/midi_driver.cpp
+++ b/core/os/midi_driver.cpp
@@ -31,7 +31,6 @@
 #include "midi_driver.h"
 
 #include "core/input/input.h"
-#include "core/os/os.h"
 
 uint8_t MIDIDriver::last_received_message = 0x00;
 MIDIDriver *MIDIDriver::singleton = nullptr;