--- a/editor/dependency_editor.h
+++ b/editor/dependency_editor.h
@@ -31,8 +31,8 @@
 #ifndef DEPENDENCY_EDITOR_H
 #define DEPENDENCY_EDITOR_H
 
-#include "editor_file_dialog.h"
 #include "scene/gui/dialogs.h"
+#include "scene/gui/item_list.h"
 #include "scene/gui/tab_container.h"
 #include "scene/gui/tree.h"