--- a/editor/dictionary_property_edit.cpp
+++ b/editor/dictionary_property_edit.cpp
@@ -29,7 +29,7 @@
 /*************************************************************************/
 
 #include "dictionary_property_edit.h"
-#include "editor_node.h"
+#include "editor/editor_node.h"
 
 void DictionaryPropertyEdit::_notif_change() {
 	notify_property_list_changed();