--- a/editor/editor_path.cpp
+++ b/editor/editor_path.cpp
@@ -30,8 +30,8 @@
 
 #include "editor_path.h"
 
-#include "editor_node.h"
-#include "editor_scale.h"
+#include "editor/editor_node.h"
+#include "editor/editor_scale.h"
 
 void EditorPath::_add_children_to_popup(Object *p_obj, int p_depth) {
 	if (p_depth > 8) {