--- a/editor/editor_plugin_settings.h
+++ b/editor/editor_plugin_settings.h
@@ -32,8 +32,8 @@
 #define EDITORPLUGINSETTINGS_H
 
 #include "core/object/undo_redo.h"
+#include "editor/editor_data.h"
 #include "editor/plugin_config_dialog.h"
-#include "editor_data.h"
 #include "property_editor.h"
 #include "scene/gui/dialogs.h"