--- a/editor/editor_run_script.h
+++ b/editor/editor_run_script.h
@@ -32,7 +32,7 @@
 #define EDITOR_RUN_SCRIPT_H
 
 #include "core/object/ref_counted.h"
-#include "editor_plugin.h"
+#include "editor/editor_plugin.h"
 class EditorNode;
 class EditorScript : public RefCounted {
 	GDCLASS(EditorScript, RefCounted);