--- a/editor/editor_sectioned_inspector.cpp
+++ b/editor/editor_sectioned_inspector.cpp
@@ -30,7 +30,7 @@
 
 #include "editor_sectioned_inspector.h"
 
-#include "editor_scale.h"
+#include "editor/editor_scale.h"
 
 class SectionedInspectorFilter : public Object {
 	GDCLASS(SectionedInspectorFilter, Object);