--- a/editor/editor_toaster.cpp
+++ b/editor/editor_toaster.cpp
@@ -28,12 +28,13 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
+#include "editor_toaster.h"
+
 #include "editor/editor_node.h"
 #include "editor/editor_scale.h"
+#include "scene/gui/button.h"
 #include "scene/gui/label.h"
 #include "scene/gui/panel_container.h"
-
-#include "editor_toaster.h"
 
 EditorToaster *EditorToaster::singleton = nullptr;