--- a/editor/export_template_manager.cpp
+++ b/editor/export_template_manager.cpp
@@ -36,8 +36,8 @@
 #include "core/io/zip_io.h"
 #include "core/os/keyboard.h"
 #include "core/version.h"
-#include "editor_node.h"
-#include "editor_scale.h"
+#include "editor/editor_node.h"
+#include "editor/editor_scale.h"
 #include "progress_dialog.h"
 #include "scene/gui/link_button.h"