--- a/editor/import/resource_importer_layered_texture.cpp
+++ b/editor/import/resource_importer_layered_texture.cpp
@@ -30,8 +30,7 @@
 
 #include "resource_importer_layered_texture.h"
 
-#include "resource_importer_texture.h"
-
+#include "core/config/project_settings.h"
 #include "core/error/error_macros.h"
 #include "core/io/config_file.h"
 #include "core/io/image_loader.h"