--- a/editor/import/scene_import_settings.h
+++ b/editor/import/scene_import_settings.h
@@ -31,8 +31,6 @@
 #ifndef SCENEIMPORTSETTINGS_H
 #define SCENEIMPORTSETTINGS_H
 
-#include "editor/editor_file_dialog.h"
-#include "editor/editor_inspector.h"
 #include "editor/import/resource_importer_scene.h"
 #include "scene/3d/camera_3d.h"
 #include "scene/3d/light_3d.h"
@@ -47,6 +45,8 @@
 #include "scene/gui/tree.h"
 #include "scene/resources/primitive_meshes.h"
 
+class EditorFileDialog;
+class EditorInspector;
 class SceneImportSettingsData;
 
 class SceneImportSettings : public ConfirmationDialog {