… editor/animation_track_editor.cpp
… editor/code_editor.cpp
… editor/editor_file_dialog.cpp
… editor/editor_inspector.cpp
… editor/editor_plugin.cpp
… editor/filesystem_dock.cpp
… editor/import_defaults_editor.cpp
… editor/import_defaults_editor.h
… editor/import_dock.cpp
… editor/inspector_dock.h
… editor/localization_editor.cpp
… editor/localization_editor.h
… editor/multi_node_edit.cpp
… editor/node_dock.cpp
… editor/plugins/abstract_polygon_2d_editor.cpp
… editor/plugins/abstract_polygon_2d_editor.h
… editor/plugins/animation_blend_space_1d_editor.cpp
… editor/plugins/animation_blend_space_1d_editor.h
… editor/plugins/animation_blend_space_2d_editor.cpp
… editor/plugins/animation_blend_space_2d_editor.h
… editor/plugins/animation_blend_tree_editor_plugin.cpp
… editor/plugins/animation_blend_tree_editor_plugin.h
… editor/plugins/animation_player_editor_plugin.cpp
… editor/plugins/animation_player_editor_plugin.h
… editor/plugins/animation_state_machine_editor.cpp
… editor/plugins/animation_state_machine_editor.h
… editor/plugins/animation_tree_editor_plugin.cpp
… editor/plugins/animation_tree_editor_plugin.h
… editor/plugins/asset_library_editor_plugin.cpp
… editor/plugins/audio_stream_editor_plugin.cpp