This PR aims to improve compilation time by forward-declaring some classes whenever possible/when it makes sense and cleaning up/moving some include statements. There is still some more room to improve, but I think this should be done in subsequent PRs as this manual process of finding optimization opportunities can be very time consuming.

Changes:
- Add forward-declarations of `EditorFileDialog` and `EditorNode` to several classes. This is already done in some cases, but not consistently.
- Remove unused includes of `file_dialog.h`
- Remove forward-declaration of Container in `flow_container.h` as it was used as a base class which can cause problems. (could be moved to a separate PR)
- Optimize includes in `EditorNode` and sort forward declarations
- Make missing relative includes (primarily `editor` includes) absolute for consistency

Some benchmarks (as of Feb 11 2022):
[Time of rebuild. before -> this PR]

Change in progress_dialog.h
**28s -> 23s (~17% reduction)**

Change in dependency_editor.h
**25s -> 23s (~8% reduction)**

Change in editor_data.h
**3:55m -> 3:01m (~23% reduction)**

Change in file_dialog.h
**3:47m -> 21s (~91% reduction)**

Change in line_edit.h
**4:37m -> 3:50m (~17% reduction)**

Changed all files above:
**4:30m -> 3:56 (~13% reduction)**