@Calinou 2022-02-04
This is a good start, nice work :slightly_smiling_face: 

Did you see https://github.com/godotengine/godot/issues/57569, by the way?