@Geometror 2022-02-05
@Calinou Thats quite interesting, will see what I can do with that :)
I am currently trying to create a dependency graph of the godot codebase to find some more optimization opportunities. Does anyone know a good tool (Windows/Linux) to render large/complex graphs and analyze them? I already have the DOT data.