@Calinou 2022-02-05
There's [include-what-you-use](https://include-what-you-use.org/) but I don't know how difficult it would be to run on a SCons codebase.