@akien-mga 2022-02-05
> There's [include-what-you-use](https://include-what-you-use.org/) but I don't know how difficult it would be to run on a SCons-using codebase.

It's not too difficult, it just needs a wrapper script. I ran it a few years ago, but it's not trivial to make good use of it as it affects the whole codebase.

When using it, a good first step would be to try to sort includes in `core` only (and anything else that requires changes due to core include changes).