@Geometror 2022-02-07
@Calinou @akien-mga 
As you said, getting it to work was straight-forward, however there are problems which makes it quite difficult to make good use of it:

- Sometimes it removes includes of generated headers
- platform specific includes are added (this can be easily fixed though)
- it fails at some complex templates (with specializations) which are declared across multiple files

Tried it in `core` but couldn't get it to work because of the last point above. It always fails at the GetTypeInfo template and I don't know why. If someone wants to have a look at it:
https://github.com/Geometror/godot/tree/iwyu-core-1

However, running it over the `editor` folder worked fine, although still requiring some manual work. See #57734 for some quick benchmarks.