@Geometror 2022-02-11
Rebased, squashed and did some further cleanup, now I think its ready for review.
Added some benchmarks to my original message above, feedback is always welcome :)