@Geometror 2022-02-12
Fixed all issues, except the ones regarding the dependency editor header as suggested. I was tempted to fix those, but there are still many places left where a similar cleanup could be done so this is indeed material for another PR :)